+++ lib/tusc/offset_request.rb
@@ -1,6 +1,5 @@
-require 'net/http'
+require_relative 'http_service'
 require_relative 'offset_response'
-require_relative '../core_ext/string/truncate'
 
 # Asks tus server for appriopriate offset
 #  for specific file, via upload_url
@@ -29,29 +28,11 @@ class TusClient::OffsetRequest
   # Retrieves offset via a HEAD request to the tus server
   # Returns the offset (in a OffsetResponse)
   def perform
-    logger.debug do
-      ['TUS HEAD',
-       request: { upload_url: upload_uri.to_s, header: headers }]
-    end
-
-    response = Net::HTTP.start(
-      upload_uri.host,
-      upload_uri.port,
-      use_ssl: upload_uri.scheme == 'https'
-    ) do |http|
-      http.head(upload_uri.path, headers)
-    end
-    received_header = response.each_key.collect { |k| { k => response.header[k] } }
-
-    logger.debug do
-      ['TUS HEAD',
-       response: {
-         status: response.code,
-         header: received_header,
-         body: response.body.to_s.truncate_middle(60)
-       }]
-    end
-
+    response = HttpService.head(
+      uri: upload_uri,
+      headers: headers,
+      logger: logger
+    )
     TusClient::OffsetResponse.new(response)
   end