+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Upcoming v0.4.6: FIX CreationResponse#body, tweak logs
 - FIX CreationResponse#body (for blank body)
+- Extract HttpService (head, patch, post)
 - CreationRequest logs body (used by some tus servers; e.g. Vimeo)
 - Log request/response vs. sending/receiving
 - #truncate uses ellipse (vs. 3 periods)